lib-builder: Update scripts and readme
- Abort applying patches when one fails.
- build_idf_libs.py and menuconfig.py warn if run in the wrong directory, instead of crashing.
- Use "reconfigure" instead of "menuconfig" to create the vanilla sdkconfig file.
- Mention new branch naming schema and which repository it applies to.

+++ software/lib-builder/menuconfig.py
@@ -27,6 +27,10 @@ if len(sys.argv) != 3:
 
 root = sys.argv[1]
 
+if not os.path.isfile(os.path.join(root, "configs", "defconfig.esp32")):
+    print("defconfig not found. Is the lib-builder root correct?")
+    sys.exit(1)
+
 shutil.move(os.path.join(root, "configs", "defconfig.esp32"), os.path.join(root, "configs", "defconfig.esp32vanilla"))
 try:
     shutil.copy(os.path.join(sys.argv[2]), os.path.join(root, "configs", "defconfig.esp32"))
